Change com name to com.umgc.summer2022 to use the better name. Rename assistant_service file
