+++ android/app/google-services.json
@@ -9,7 +9,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:872576625259:android:accd56b745157ebc7f1f2e",
         "android_client_info": {
-          "package_name": "com.example.summer2022"
+          "package_name": "com.umgc.summer2022"
         }
       },
       "oauth_client": [
@@ -17,7 +17,7 @@
           "client_id": "872576625259-9tvmkki1cm62g55ec4ch29d9r949bftt.apps.googleusercontent.com",
           "client_type": 1,
           "android_info": {
-            "package_name": "com.example.summer2022",
+            "package_name": "com.umgc.summer2022",
             "certificate_hash": "ad5d5f74403d7ffe8ea81cf4441ad0dd162cf5d5"
           }
         },
@@ -42,7 +42,7 @@
               "client_id": "872576625259-h9l5shh21q8cccq2lbovc8e8eq9k4l7c.apps.googleusercontent.com",
               "client_type": 2,
               "ios_info": {
-                "bundle_id": "com.example.summer2022"
+                "bundle_id": "com.umgc.summer2022"
               }
             }
           ]
