+++ android/app/src/main/kotlin/com/example/usps_informed_delivery_backend/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.summer2022
+package com.umgc.summer2022
 
 import io.flutter.embedding.android.FlutterActivity
