+++ android/app/src/main/res/xml/shortcuts.xml
@@ -12,8 +12,9 @@
     </shortcut>
     <capability android:name="actions.intent.GET_THING">
         <intent
-            android:targetPackage="com.example.summer2022"
-            android:targetClass="com.example.summer2022.MainActivity">
+            android:targetPackage="com.umgc.summer2022"
+            android:targetClass="com.umgc.summer2022.MainActivity"
+            android:action="actions.intent.GET_THING">
             <parameter
                 android:name="thing.name"
                 android:key="name"/>
@@ -21,8 +22,9 @@
     </capability>
     <capability android:name="actions.intent.CREATE_THING">
         <intent
-            android:targetPackage="com.example.summer2022"
-            android:targetClass="com.example.summer2022.MainActivity">
+            android:targetPackage="com.umgc.summer2022"
+            android:targetClass="com.umgc.summer2022.MainActivity"
+            android:action="actions.intent.CREATE_THING">
             <parameter
                 android:name="thing.name"
                 android:key="name"/>
@@ -30,8 +32,9 @@
     </capability>
     <capability android:name="actions.intent.OPEN_APP_FEATURE">
         <intent
-            android:targetPackage="com.example.summer2022"
-            android:targetClass="com.example.summer2022.MainActivity">
+            android:targetPackage="com.umgc.summer2022"
+            android:targetClass="com.umgc.summer2022.MainActivity"
+            android:action="actions.intent.OPEN_APP_FEATURE">
             <parameter
                 android:name="feature"
                 android:key="name"/>
