+++ lib/main.dart
@@ -3,7 +3,7 @@
 import 'package:flutter/material.dart';
 import 'package:global_configuration/global_configuration.dart';
 import 'package:loader_overlay/loader_overlay.dart';
-import 'package:summer2022/services/assistantService.dart';
+import 'package:summer2022/services/assistant_Service.dart';
 import 'package:summer2022/services/cache_service.dart';
 import 'package:summer2022/services/mail_utility.dart';
 import 'package:summer2022/services/background_service.dart';
